almost final version

+++ includes/vector_operations.h
@@ -6,7 +6,7 @@
 /*   By: meldora <meldora@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/15 18:08:29 by meldora           #+#    #+#             */
-/*   Updated: 2021/01/31 15:50:28 by meldora          ###   ########.fr       */
+/*   Updated: 2021/03/02 16:31:20 by meldora          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,10 +16,10 @@
 # include <math.h>
 # include <stdlib.h>
 
-float	get_length(float v[3]);
-void	normalize(float v[3]);
-float	dot(float u[3], float v[3]);
-float	*cross(float u[3], float v[3]);
-float	*get_normalized(float v[3]);
+double	get_length(double v[3]);
+void	normalize(double v[3]);
+double	dot(double u[3], double v[3]);
+double	*cross(double u[3], double v[3]);
+double	*get_normalized(double v[3]);
 
 #endif
